[JENKINS-34935] Fixing detection of corrupted zip files
This commits changes the test for a broken zip file to a full CRC test
of every entry of the zip file.

This will ensure that zip files with even marginal errors inside the
different zip entries are detected properly.

Added also two tests to verify that the task properly detects broken
zip files and zip files that are okay.