use pinned version for commons-lang3-api
It would be nice to use a pinned version for the dependency commons-lang3-api. My value is the latest version, as listed here: https://plugins.jenkins.io/commons-lang3-api/
Dependabot should be able to raise the version in the future. We are currently having problems with dependencies of other Jenkins plugins that use version 3.12.0-36.vd97de6465d5b_ and commons-lang3-api forcing 3.12.0.0